Backend/requirements.txt
a==1.0
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.3
cryptography==37.0.4
cssselect2==0.6.0
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.5
dj-static==0.0.6
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==3.3
install==1.3.5
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.8.1
jsonschema==4.9.1
lxml==4.9.1
more-itertools==8.13.0
of==1.0.1
oscrypto==1.3.0
packaging==21.3
Pillow==9.2.0
portend==3.1.0
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
PyPDF2==2.10.0
PyPDF3==1.0.6
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
svglib==1.4.1
tempora==5.0.2
tinycss2==1.1.1
to==0.3
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.11
webencodings==0.5.1
wrapt==1.14.1
ws4py==0.5.1
xhtml2pdf==0.2.8
zc.lockfile==2.0
